Check for Atom namespace when resolve content type
